move the query specific funcionality to out of the project view camponent and into the queries component
