+++ client/components/ChartBox.tsx
@@ -134,8 +134,6 @@ const ChartBox: React.FC<IProps> = ({ queries }) => {
                 label: 'Tokens',
                 borderColor: 'rgba(255, 99, 132, 0.5)',
                 backgroundColor: 'rgba(255, 99, 132, 0.5)',
-                borderWidth: 2,
-                fill: false,
                 data: tokenData,
                 tension: 0.5,
                 pointStyle: 'line',
@@ -167,7 +165,6 @@ const ChartBox: React.FC<IProps> = ({ queries }) => {
                 backgroundColor: 'rgba(75, 192, 192, 0.8)',
                 data: depthData,
                 borderColor: 'rgba(75, 192, 192, 0.8)',
-                borderWidth: 2,
                 tension: 0.5,
                 pointStyle: 'line',
             },
