+++ client/components/Queries.tsx
@@ -1,79 +1,154 @@
-import React from 'react';
-// import Collapsible from 'react-collapsible';
-// import { Projects, Project, ProjectQuery } from './Interfaces';
+import React, { useState } from 'react';
 import { ProjectQuery } from '../../@types/Interfaces';
 import { SortOrder } from '../../@types/dashboard';
 import Query from './Query';
 
-export interface IProps {
-    // projects: Projects['projects'];
-    // test: any;
-    // combinedSort: (field: 'timestamp' | 'depth' | 'complexity', sortOrder: SortOrder) => void;
-    // arrowTime: string | undefined;
-    // arrowDepth: string | undefined;
-    // arrowComplexity: string | undefined;
-    // setToggle: any;
-    // time: boolean;
-    // depth: boolean;
-    // complexity: boolean;
-    // sortByTime: any;
-    // sortByDepth: any;
-    // sortByComplexity: any;
-    queries: ProjectQuery[];
+export interface ISState {
+    style: {
+        time: boolean; // TODO: Change to timestamp
+        depth: boolean;
+        complexity: boolean;
+    };
+    arrow: {
+        timestamp: SortOrder;
+        depth: SortOrder;
+        complexity: SortOrder;
+    };
 }
+
+export interface IProps {
+    rawQueries: ProjectQuery[];
+}
+
 // eslint-disable-next-line react/function-component-definition
-const Queries: React.FC<IProps> = ({
-    // combinedSort,
-    // arrowTime,
-    // arrowDepth,
-    // arrowComplexity,
-    // setToggle,
-    // time,
-    // depth,
-    // complexity,
-    // sortByTime,
-    // sortByDepth,
-    // sortByComplexity,
-    queries,
-}) => (
-    <div id="loggerBtnWrapper">
-        {/* -------this is left for the future project implementation. Waiting for instruction
-                {projects?.map((project: Project) => (
-                <div aria-hidden="true" onClick={() => test(project)}>
-                    <Collapsible
-                        trigger={<span style={{ padding: '10px 50px' }}>{project.name}</span>}
-                        className="projectCard"
+const Queries: React.FC<IProps> = ({ rawQueries }) => {
+    const [style, setStyle] = useState<ISState['style']>({
+        time: false,
+        depth: false,
+        complexity: false,
+    });
+    const [arrow, setArrow] = useState<ISState['arrow']>({
+        // adjusting arrow based on ascending vs descending
+        timestamp: '',
+        depth: '',
+        complexity: '',
+    });
+
+    const [listOfQueries, setListOfQueries] = useState(rawQueries);
+
+    const combinedSort = (field: keyof ISState['arrow'], sortOrder: SortOrder): void => {
+        const newArr = [...listOfQueries];
+        newArr.sort((a, b) => {
+            if (sortOrder === '↑') {
+                return a[field] - b[field];
+            }
+            if (sortOrder === '↓') {
+                return b[field] - a[field];
+            }
+            return 0;
+        });
+        setArrow({
+            timestamp: '',
+            depth: '',
+            complexity: '',
+            [field]: sortOrder,
+        });
+        setListOfQueries(newArr);
+    };
+
+    const setToggle = (arg: string) => {
+        if (arg === 'time') {
+            setStyle({
+                ...style,
+                time: true,
+                depth: false,
+                complexity: false,
+            });
+        } else if (arg === 'depth') {
+            setStyle({
+                ...style,
+                time: false,
+                depth: true,
+                complexity: false,
+            });
+        } else if (arg === 'complexity') {
+            setStyle({
+                ...style,
+                time: false,
+                depth: false,
+                complexity: true,
+            });
+        }
+    };
+
+    return (
+        <>
+            <div className="loggerSortButtonsWrapper">
+                <div className="loggerSortButtons">
+                    <div
+                        aria-hidden="true"
+                        className={`loggerBtn${style.time ? ' active' : ''}`}
+                        onClick={() => {
+                            // sortByTime();
+                            let sortOrder: SortOrder;
+                            if (arrow.timestamp === '' || arrow.timestamp === '↓') {
+                                sortOrder = '↑';
+                            } else {
+                                sortOrder = '↓';
+                            }
+
+                            combinedSort('timestamp', sortOrder);
+                            setToggle('time');
+                        }}
                     >
-                        <div>
-                            {project.queries.map((query: ProjectQuery) => (
-                                <div className="queryProps">
-                                    <div>
-                                        <div className="label">Name: </div>
-                                        <div className="value">{query.name}</div>
-                                    </div>
-                                    <div>
-                                        <div className="label">Time: </div>
-                                        <div className="value">{query.time}</div>
-                                    </div>
-                                    <div>
-                                        <div className="label">Depth: </div>
-                                        <div className="value">{query.depth}</div>
-                                    </div>
-                                    <div>
-                                        <div className="label">Complexity: </div>
-                                        <div className="value">{query.complexity}</div>
-                                    </div>
-                                </div>
-                            ))}
-                        </div>
-                    </Collapsible>
+                        Time {arrow.timestamp}
+                    </div>
+                    <div
+                        aria-hidden="true"
+                        className={`loggerBtn${style.depth ? ' active' : ''}`}
+                        onClick={() => {
+                            // sortByDepth();
+                            let sortOrder: SortOrder;
+                            if (arrow.depth === '' || arrow.depth === '↓') {
+                                sortOrder = '↑';
+                            } else {
+                                sortOrder = '↓';
+                            }
+                            combinedSort('depth', sortOrder);
+                            setToggle('depth');
+                        }}
+                    >
+                        Depth {arrow.depth}
+                    </div>
+                    <div
+                        aria-hidden="true"
+                        className={`loggerBtn${arrow.complexity ? ' active' : ''}`}
+                        onClick={() => {
+                            // sortByComplexity();
+                            let sortOrder: SortOrder;
+                            if (arrow.complexity === '' || arrow.complexity === '↓') {
+                                sortOrder = '↑';
+                            } else {
+                                sortOrder = '↓';
+                            }
+                            combinedSort('complexity', sortOrder);
+                            setToggle('complexity');
+                        }}
+                    >
+                        Complexity {arrow.complexity}
+                    </div>
                 </div>
-            ))} */}
-        <div className="space" />
-        {queries?.map((query: ProjectQuery) => (
-            <Query query={query} />
-        ))}
-    </div>
-);
+            </div>
+            <div className="loggerGUI">
+                <div id="loggerBtnWrapper">
+                    <div className="space" />
+                    {listOfQueries?.map((query: ProjectQuery) => (
+                        <Query query={query} />
+                    ))}
+                </div>
+            </div>
+        </>
+    );
+};
 
 export default Queries;
