+++ public/styles.css
@@ -501,11 +501,13 @@ section#pageThree {
     border-bottom: 0px;
 }
 .label {
+    font-size: 10px;
     font-weight: bold;
     display: inline-block;
     margin-right: 10px;
 }
 .value {
+    font-size: 10px;
     text-align: center;
     display: inline-block;
     margin-right: 25px;
